discounts

+++ database/migrations/2025_02_26_162736_create_pedido_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pedido', function (Blueprint $table) {
             $table->id();
-            $table->integer('cost');
+            $table->float('cost');
             $table->foreignId('food_id')->nullable()->references('id')->on('food');
